+++ ron-test/LwwStruct.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes #-}
 
@@ -28,10 +29,10 @@ import           String (s)
 
 example0 :: Struct51
 example0 = Struct51
-    { int1 = 275
-    , str2 = RGA "275"
-    , str3 = "190"
-    , set4 = ORSet []
+    { int1 = Just 275
+    , str2 = Just $ RGA "275"
+    , str3 = Just "190"
+    , set4 = Just $ ORSet []
     , opt5 = Nothing
     , opt6 = Just 74
     }
@@ -45,7 +46,7 @@ ex1expect = [s|
     *lww    #B/0000000DrW+r3pl1c4                   !
                                     @`      :int1   275
                                             :opt5
-                                            :opt6   >some 74
+                                            :opt6   74
                                             :set4   >}KUW
                                             :str2   >}OUW
                                             :str3   '190'
@@ -98,17 +99,18 @@ ex4expect = [s|
 
 example4expect :: Struct51
 example4expect = Struct51
-    { int1 = 166
-    , str2 = RGA "145"
-    , str3 = "206"
-    , set4 = ORSet [Struct51
-        { int1 = 135
-        , str2 = RGA "136"
-        , str3 = "137"
-        , set4 = ORSet []
-        , opt5 = Nothing
-        , opt6 = Nothing
-        }]
+    { int1 = Just 166
+    , str2 = Just $ RGA "145"
+    , str3 = Just "206"
+    , set4 = Just $ ORSet
+        [Struct51
+            { int1 = Just 135
+            , str2 = Just $ RGA "136"
+            , str3 = Just "137"
+            , set4 = Just $ ORSet []
+            , opt5 = Nothing
+            , opt6 = Nothing
+            }]
     , opt5 = Nothing
     , opt6 = Nothing
     }
@@ -135,18 +137,18 @@ prop_lwwStruct = property $ do
         runNetworkSimT $ runReplicaSimT replica $
         execObjectState ex2state $ do
             -- plain field
-            int1_assign 166
+            int1_assign $ Just 166
             str2_zoom $ RGA.edit "145"
             str3Value <- str3_read
-            str3Value === "190"
-            str3_assign "206"
+            str3Value === Just "190"
+            str3_assign $ Just "206"
             set4_zoom $
                 ORSet.addValue
                     Struct51
-                        { int1 = 135
-                        , str2 = RGA "136"
-                        , str3 = "137"
-                        , set4 = ORSet []
+                        { int1 = Just 135
+                        , str2 = Just $ RGA "136"
+                        , str3 = Just "137"
+                        , set4 = Just $ ORSet []
                         , opt5 = Nothing
                         , opt6 = Nothing
                         }