+++ ron-test/StructSet.hs
@@ -38,7 +38,7 @@ example0 = StructSet13
     , str2 = Just $ RGA "275"
     , str3 = Just "190"
     , set4 = Just $ ORSet []
-    , opt5 = Just Nothing
+    , opt5 = Nothing
     , nst6 = Nothing
     }
 
@@ -50,14 +50,15 @@ state1expect :: ByteStringL
 state1expect = [s|
     *set    #B/0000000DrW+r3pl1c4           !
                                     @`}KUW  >int1   275
-                                    @}OUW   >opt5
-                                    @}WUW   >set4   >B/0000000lUW+r3pl1c4
-                                    @{10UW  >str2   >B/0000001GUW+r3pl1c4
-                                    @}LcW   >str3   '190'
+                                    @}OUW   >nst6
+                                    @}WUW   >opt5
+                                    @}lUW   >set4   >B/00000010UW+r3pl1c4
+                                    @{1GUW  >str2   >B/0000001WUW+r3pl1c4
+                                    @}acW   >str3   '190'
 
-            #}lUW                   @0      !
+            #{10UW                  @0      !
 
-    *rga    #{1GUW                          !
+    *rga    #}WUW                           !
                                     @`]g6   '2'
                                     @)7     '7'
                                     @)8     '5'
@@ -67,32 +68,41 @@ state1expect = [s|
 state4expect :: ByteStringL
 state4expect = [s|
     *set    #B/0000000DrW+r3pl1c4                   !
-                                    @`}OUW          >opt5
-                                    @}WUW           >set4 >B/0000000lUW+r3pl1c4
-                                    @{10UW          >str2 >B/0000001GUW+r3pl1c4
-                                    @}ZdW           >int1 166
-                                    @}k2W   :`{0KUW >int1 275
-                                    @{2OUW  :0      >str3 '206'
-                                    @}WUW   :`{1LcW >str3 '190'
-                                    @{3~2W  :0      >nst6 >B/00000042MW+r3pl1c4
-            #}lUW                   @0              !
-                                    @`{3odW         >{2lUW
-    *rga    #{1GUW                  @0              !
-                                    @`]g6   :`}nMW  '2'
-                                    @)7     :{21UW  '7'
-                                    @{2AUW  :0      '1'
-                                    @}KUW           '4'
+                                    @`}WUW          >opt5
+                                    @}lUW           >set4 >B/00000010UW+r3pl1c4
+                                    @{1GUW          >str2 >B/0000001WUW+r3pl1c4
+                                    @}odW           >int1 166
+                                    @}~2W   :`{0KUW >int1 275
+                                    @{2dUW  :0      >str3 '206'
+                                    @}lUW   :`{1acW >str3 '190'
+                                    @{4~2W  :0      >nst6 >B/00000052MW+r3pl1c4
+                                    @{6GUW  :`{0OUW >nst6
+            #{10UW                  @0      :0      !
+                                    @`{4odW         >{30UW
+    *rga    #}WUW                   @0              !
+                                    @`]g6   :`{22MW '2'
+                                    @)7     :}HUW   '7'
+                                    @{2QUW  :0      '1'
+                                    @}_UW           '4'
                                     @`]g8           '5'
-    *set    #{2lUW                  @0              !
-                                    @`{30UW         >int1 135
-                                    @}GUW           >str2 >B/0000003WUW+r3pl1c4
+    *set    #{30UW                  @0              !
+                                    @`}GUW          >int1 135
+                                    @}WUW           >nst6
+                                    @}lUW           >opt5
+                                    @{40UW          >set4
+                                    @}GUW           >str2 >B/0000004WUW+r3pl1c4
                                     @}acW           >str3 '137'
-    *rga    #{3WUW                  @0              !
+    *rga    #{4WUW                  @0              !
                                     @`]g6           '1'
                                     @)7             '3'
                                     @)8             '6'
-    *set    #{42MW                  @0              !
+    *set    #{52MW                  @0              !
                                     @`}HUW          >int1 138
+                                    @}QUW           >nst6
+                                    @}_UW           >opt5
+                                    @}dUW           >set4
+                                    @}lUW           >str2
+                                    @{60UW          >str3
     .
     |]
 
@@ -103,7 +113,7 @@ example4expect = StructSet13
     , str3 = Just "206"
     , set4 = Just $
         ORSet [def{int1 = Just 135, str2 = Just $ RGA "136", str3 = Just "137"}]
-    , opt5 = Just Nothing
+    , opt5 = Nothing
     , nst6 = Just def{int1 = Just 138}
     }
 
@@ -145,7 +155,7 @@ prop_structSet = property $ do
                 checkCausality
             checkCausality
             do  value <- opt5_read
-                value === Just Nothing
+                value === Nothing
             do  value <- nst6_read
                 value === Nothing
             nst6_assign def{int1 = Just 138}